#523 move params update from worker
